# [Discussion] Cull actors outside of AI processing range