--- a/apps/openmw/mwphysics/physicssystem.hpp
+++ b/apps/openmw/mwphysics/physicssystem.hpp
@@ -86,6 +86,8 @@ namespace MWPhysics
             void removeHeightField (int x, int y);
 
             bool toggleCollisionMode();
+            bool isActorCollisionEnabled(const MWWorld::Ptr& ptr);
+            void setActorCollisionMode(const MWWorld::Ptr& ptr, bool enabled);
 
             void stepSimulation(float dt);
             void debugDraw();