--- a/files/settings-default.cfg
+++ b/files/settings-default.cfg
@@ -200,6 +200,9 @@ best attack = false
 # Difficulty.  Expressed as damage dealt and received. (e.g. -100 to 100).
 difficulty = 0
 
+# The maximum range of actor AI, animations and physics updates.
+actors processing range = 7168
+
 # Make reflected Absorb spells have no practical effect, like in Morrowind.
 classic reflected absorb spells behavior = true