… apps/openmw/mwworld/worldimp.cpp
… docs/source/reference/modding/settings/game.rst
… files/mygui/openmw_settings_window.layout