Related to [bug #4647](https://gitlab.com/OpenMW/openmw/issues/4647).
Do we really need to render all actors in scene indifferently from distance, and for flying creatures update physics too?

I tried to do not render actors outside of AI processing range, and here is result of my experiments.
Comparison in Caldera on i7-7700:

exterior cell load distance = 3:
Master: 25 FPS
PR:55 FPS

exterior cell load distance = 2:
Master: 55 FPS
PR: 75 FPS

exterior cell load distance = 1:
Master: 100 FPS
PR: 110 FPS

Can anyone test how it affects performance?
Also I wonder if we can apply fog for actors near processing border.
@AnyOldName3, @wareya, what do you think?