@AnyOldName3 2018-09-23
Fog would look silly - the actors wouldn't be any less visible, but would look like white ghostly silhouettes. I'm pretty sure that the later games make things semi-transparent (alpha-blended) but then that forces all semi-distant actors into the depth-sorted translucent render bin, which is slower. In actual fact, that shouldn't add any more cost than one tree per sixteen actor drawables (I don't know how many drawables make up an actor usually) so if you're seeing huge speedups as-is, it should still be a net improvement.