@akortunov 2018-09-24
> I'm pretty sure that the later games make things semi-transparent (alpha-blended) but then that forces all semi-distant actors into the depth-sorted translucent render bin, which is slower. 

How can I apply transparency to actor? In the similar way as we handle Chameleon effect?