@akortunov 2018-09-24
> No idea, but something like that would be where I'd start looking.

Animation::setAlpha() does the trick. As for me, result is a quite good.