@wareya 2018-09-24
What happens when an actor that's fading away is half in water? Do they render before or after the water? OpenMW already has some alpha-sorting bugs (images attached), it would be annoying to add more.

https://user-images.githubusercontent.com/585488/45955496-14f44d80-bfde-11e8-92f3-b0b88602b109.png
https://user-images.githubusercontent.com/585488/45955497-14f44d80-bfde-11e8-91ac-70b551f580f4.png
https://user-images.githubusercontent.com/585488/45955498-14f44d80-bfde-11e8-908f-661b857cf6b0.png