@kcat 2018-09-24
If possible, it's best to avoid alpha transparency. Some kind of dither or stipple "transparency" could work, which simply discards a percentage of rendered pixels and leaves the remaining ones opaque, avoiding the sorting issues. I don't know exactly how you'd set that up, though.