@akortunov 2018-09-24
> What happens when an actor that's fading away is half in water?

The pretty same thing if you will cast invisibility and go underwater.
Also keep in mind that water is less transparent on large distance.