@akortunov 2018-09-24
Added basic implementation of fading.
Just a proof of concept! Should be enough to check if it causes issues.
I like the idea of dithered transparency, but I do not know how to implement it yet.