@akortunov 2018-09-24
Unfortunately, I am complely unfamiliar with shaders, so I doubt I can do it with the "sane amount of effort".
I suggest to test this PR for now and check if it ever worth efforts to implement a dither shader.