@akortunov 2018-09-24
About shader: I suppose, there should be something like [this](https://gist.github.com/jmptable/eec7b62444aa92bcb33a), right?