@akortunov 2018-10-02
@zinnschlag, what do you think? Is the dithering shader a blocker?