@wareya 2018-10-03
As one of the people who recommended using dithering instead of alpha, it's not a blocker, it's just something that would be nice to fix before 1.0 if it goes in with alpha.