@akortunov 2018-10-06
An another question, which I want to discuss: should we implement an actor provessing distance slider?
Currently we have a hardcoded distance 7168, and we can use it as a cap (since there are issues with larger values).
Since out "actor processing" includes not only AI, but physics stuff too, configurable option can allow us to increase performance on CPUs with low per-core performance.

What do you think.