@Capostrophic 2018-10-06
+1. Though I'm not sure which tab should such a slider belong to.