@akortunov 2018-10-06
Added expiremental support of processing range slider.
With lowest value it has 1/2 of our current processing range, so processing square will be 4 times smaller.
That should help with performance drops in cities on low-performance machines.
Not sure which changes should be done to make it work with the TES3MP.
@davidcernat, what do you think?