@akortunov 2018-10-07
In the Animation::setAlpha() method we re-create shaders for actor every time we change transparency a bit. Probably we need a better way to fade actors.