@zinnschlag 2018-10-07
The actor processing slider is a bit risky. Some content could break is getting too low. I guess we can keep it for now and later introduce settings in the content file format that define the allowed ranges for actor processing.