@Capostrophic 2018-10-07
If the caption of the slider stays this way, I guess "AI Distance" GMST would become redundant to be ever included.