@AnyOldName3 2018-10-07
> In the Animation::setAlpha() method we re-create shaders for actor every time we change transparency a bit. Probably we need a better way to fade actors.
> 
> As I understood, the only reason why we re-create shaders here, is because we create a material every time, right?

Can we edit the existing material (provided the actor has a unique copy)? I'm pretty sure that the material doesn't actually affect shader preprocessing or compilation, so it should be fine to reuse it.