@akortunov 2018-10-07
I suppose scrawl added this code for reason: 268594dcf145b17c98c4def32d78c334392d593b