@akortunov 2018-10-24
I suppose this PR is ready for review and testing.