@zinnschlag 2018-10-25
Source looks okay to me (still have to do a test-run myself). @AnyOldName3 @Capostrophic Okay with this?