@Capostrophic 2018-10-25
"Actors processing distance" is not capitalized like it should ( (like Actors Processing Distance) on English version which is supposed to be the priority, and it looks out of place on other localizations anyways. Maybe resort to sAIDistance GMST as the caption of the slider?

Actor fading and everything else seems to work OK.