@akortunov 2018-10-25
> Maybe resort to sAIDistance GMST as the caption of the slider?

As I said earlier, this setting is not an AI distance. "Processing" here means a full update: rendering, physics, animations, etc.