@Capostrophic 2018-10-25
> As I said earlier, this setting is not an AI distance.

I **do not** argue that it's not simply AI simulation distance. But it's not *that* misleading from the user's standpoint, and while there are no mechanics in place to support GUI localization the current caption distracts much more than an universally available string that at least somewhat serves the purpose.